Moved backend json data into separate file and called function to store.
