+++ .gitignore
@@ -1,0 +1,3 @@
+# Visual studio configs
+.vscode/launch.json
+
