+++ server/data/attendance.json
@@ -2,32 +2,32 @@
     {
         "id": 123,
         "classBeginDate": "2021-10-01",
-        "dayAttended": 45
+        "daysAttended": 45
     },
     {
         "id": 234,
         "classBeginDate": "2021-10-01",
-        "dayAttended": 30
+        "daysAttended": 30
     },
     {
         "id": 345,
         "classBeginDate": "2021-10-01",
-        "dayAttended": 22
+        "daysAttended": 22
     },
     {
         "id": 567,
         "classBeginDate": "2021-10-01",
-        "dayAttended": 40
+        "daysAttended": 40
     },
     {
         "id": 678,
         "classBeginDate": "2021-10-01",
-        "dayAttended": 10
+        "daysAttended": 10
     },
     {
         "id": 789,
         "classBeginDate": "2021-10-01",
-        "dayAttended": 43
+        "daysAttended": 43
     }
 
 ]
